requirements.txt
numpy
opencv-python
pandas
pillow
scikit-image
scikit-learn
tensorflow
tqdm
